Finished OpenFile, DataObject

+++ src/main/java/shaggydev/models/RowString.java
@@ -20,27 +20,27 @@ public class RowString {
         return _list;
     }
 
-    public List<Float> getRow(){
+    public List<Double> getRow(){
 
-        List<Float> dlist = new ArrayList<>();
+        List<Double> dlist = new ArrayList<>();
         for(int i=0;i<_list.size();i++) {
             try {
-                dlist.add(Float.parseFloat(_list.get(i)));
+                dlist.add(Double.parseDouble(_list.get(i)));
             }catch (Exception e) {
-                dlist.add(0f);
+                dlist.add(0.0);
             }
         }
         return dlist;
     }
 
-    public List<Float> getRow(int col_title){
-        List<Float> dlist = new ArrayList<>();
+    public List<Double> getRow(int col_title){
+        List<Double> dlist = new ArrayList<>();
         for(int i=0;i<_list.size();i++) {
             if(i!=col_title) {
                 try {
-                    dlist.add(Float.parseFloat(_list.get(i)));
+                    dlist.add(Double.parseDouble(_list.get(i)));
                 }catch (Exception e) {
-                    dlist.add(0f);
+                    dlist.add(0.0);
                 }
             }
         }
